move RunAllProofsInfFlow.java
