+++ key.core.infflow/src/main/resources/META-INF/services/de.uka.ilkd.key.speclang.infflow.InformationFlowContractSupplier
@@ -1,0 +1,1 @@
+de.uka.ilkd.key.informationflow.impl.InformationFlowContractImplSupplier
