+++ key.core.infflow/src/test/java/de/uka/ilkd/key/proof/runallproofs/RunAllProofsInfFlow.java
@@ -4,6 +4,8 @@
 package de.uka.ilkd.key.proof.runallproofs;
 
 import java.io.IOException;
+import java.nio.file.Paths;
+import java.util.List;
 import java.util.stream.Stream;
 
 import de.uka.ilkd.key.proof.runallproofs.proofcollection.StatisticsFile;
@@ -12,12 +14,6 @@ import org.junit.jupiter.api.*;
 
 /**
  * This test case captures all information flow run-all-proof scenarios.
- * <p>
- * The test case is controlled by the index file (see {@value #INDEX_FILE}).
- * <p>
- * If the property "{@value #SKIP_INF_FLOW_PROPERTY}" is set to true, then no
- * info-flow
- * run-all-proof tests will be run.
  *
  * @author M. Ulbrich
  */
@@ -27,10 +23,15 @@ import org.junit.jupiter.api.*;
 public final class RunAllProofsInfFlow {
     @TestFactory
     Stream<DynamicTest> data() throws IOException {
-        var proofCollection = ProofCollections.automaticInfFlow();
+        var proofCollection = InfFlowProofCollection.automaticInfFlow();
         StatisticsFile statisticsFile = proofCollection.getSettings().getStatisticsFile();
         statisticsFile.setUp();
         Assumptions.assumeTrue(proofCollection != null);
         return RunAllProofsTest.data(proofCollection);
+    }
+
+    public static void main(String[] args) throws IOException {
+        GenerateUnitTests.run(Paths.get("build/generated-src/"),
+            List.of(InfFlowProofCollection.automaticInfFlow()));
     }
 }
