+++ key.core/build.gradle
@@ -131,15 +131,6 @@ tasks.register('testRunAllFunProofs', Test) {
     }
 }
 
-tasks.register('testRunAllInfProofs', Test) {
-    description = 'Prove/reload all keyfiles tagged for regression testing'
-    group = "verification"
-    filter {
-        includeTestsMatching "RunAllProofsInfFlow"
-    }
-}
-
-
 tasks.register('testProveSMTLemmas', Test) {
     description = 'Prove (or load proofs for) lemmas used in the SMT translation'
     group = "verification"
