+++ key.ui/examples/InformationFlow/ToyVoting/Voter(Voter__publishVoterParticipation()).Non-interference contract.0.key
@@ -1,4 +1,4 @@
-\profile "Java Profile";
+\profile "java-infflow";
 
 \settings {
 "#Proof-Settings-Config-File
