+++ key.ui/examples/InformationFlow/ToyVoting/Voter(Voter__secure_voting()).Non-interference contract.0.m.key
@@ -1,4 +1,4 @@
-\profile "Java Profile";
+\profile "java-infflow";
 
 \settings {
 "#Proof-Settings-Config-File
