+++ key.ui/examples/InformationFlow/ToyVoting/project.key
@@ -1,6 +1,6 @@
 // project file
 
-\profile "Java Profile";
+\profile "java-infflow";
 
 \settings {
 "#Proof-Settings-Config-File
